Commit 1
I import on line 1 javax.swing.* to use Swing components like JFrame, JPanel, and JButton.
I import on line 2, java.awt.** to use layouts like CardLayout and BorderLayout.
I declare on line 4,  the public class MenuPanelApp { to build the program.
Inside it,
 I declare a public JFrame named frame as the main application window.
I declare on lines 5 and 6 a public JPanel named main that will serve as the container for multiple panels.
 I declare on line 7,  a public CardLayout named card to manage switching between those panels.
I define on line 8, a public constructor MenuPanelApp() { to initialize the GUI when the object is created.