Commit 2
I added an ActionListener on line 35,  an ActionListener to the home, so it switches to the 'Home' panel when clicked.
 I added on line 36, an ActionListener to the profile so it switches to the 'Profile' panel when clicked.
I added on line 37  to the settings so it switches to the 'Settings' panel when clicked.
 I added on line 38, an ActionListener to exitItem so the program exits when clicked.
 I added on line 39, the main (the CardLayout panel container) to the frame.
I set on line 40,  the frame visible with setVisible(true) to show the GUI.
I close on line 41  the constructor with }.